requirements.txt
cachelib==0.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
docopt==0.6.2
Flask==2.0.2
Flask-Session==0.4.0
idna==3.3
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
numpy==1.21.3
Pillow==8.4.0
pipreqs==0.4.11
requests==2.26.0
torch==1.10.0
torchvision==0.11.1
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
yarg==0.1.9
zipp==3.6.0
